+++ e2e/commands/tag.e2e.1.ts
@@ -603,7 +603,7 @@ describe('bit tag command', function() {
     it('should not tag and throw an error regarding the relative syntax', () => {
       expect(output).to.have.string('error: issues found with the following component dependencies');
       expect(output).to.have.string(
-        'components with relative import statements (please use absolute paths for imported components)'
+        'components with relative import statements (please use module paths for imported components)'
       );
       expect(output).to.have.string(`${helper.scopes.remote}/utils/is-type@0.0.1`);
     });
@@ -962,7 +962,7 @@ describe('bit tag command', function() {
       helper.fs.createFile('utils', 'is-string.js', fixtures.isString);
       helper.command.addComponent('utils/is-type.js');
       helper.command.addComponent('utils/is-string.js');
-      output = helper.general.runWithTryCatch('bit tag is-string');
+      output = helper.general.runWithTryCatch('bit tag is-string --allow-relative-paths');
     });
     it('should show a descriptive error message', () => {
       expect(output).to.have.string('this dependency was not included in the tag command');