+++ e2e/functionalities/workspace-config.e2e.3.ts
@@ -399,6 +399,7 @@ describe('workspace config', function() {
           );
 
           // an intermediate step, make sure bit status shows the component with an issue of a missing file
+          helper.command.linkAndRewire();
           const status = helper.command.status();
           expect(status).to.have.string(statusFailureMsg);
 
@@ -446,6 +447,7 @@ describe('workspace config', function() {
           showBar = helper.command.showComponentParsed('bar');
         });
         it('bit status should not show the component as missing component', () => {
+          helper.command.linkAndRewire();
           const status = helper.command.status();
           expect(status).to.not.have.string(statusFailureMsg);
         });