+++ src/consumer/component-ops/install-components.ts
@@ -20,7 +20,8 @@ export async function install(consumer: Consumer, verbose: boolean): Promise<Lin
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   const dirsAbsolute = dirs.map(dir => path.join(consumerPath, dir));
   await installPackages(consumer, dirsAbsolute, verbose, true);
-  return linkAllToNodeModules(consumer);
+  const { linksResults } = await linkAllToNodeModules(consumer);
+  return linksResults;
 }
 
 /**